refactor: move out functions
